requirements-hf.txt
# ML Study Buddy - Hugging Face Spaces Requirements

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# LangChain & RAG
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.0
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.2.0

# Vector Store & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# ML Models
transformers>=4.35.0
torch>=2.0.0
torchaudio>=2.0.0
scipy>=1.11.0
datasets>=2.14.0

# Document Processing
PyMuPDF>=1.23.0
beautifulsoup4>=4.12.0
requests>=2.31.0
Pillow>=10.0.0

# Environment
python-dotenv>=1.0.0

# Audio Processing (for Whisper)
librosa>=0.10.0
soundfile>=0.12.0
